Updates
-Added Authorozation method for easy coding
