+++ pages/api/post/UserManagement/UpdateUser/index.ts
@@ -1,0 +1,12 @@
+import { NextApiRequest, NextApiResponse } from "next";
+import authorizationHandler from "pages/api/authorizationHandler";
+
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  const authorized = await authorizationHandler("POST", req, res);
+  if (authorized) {
+    return false;
+  }
+}
